Stage 1 sound refactor
